% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dates_na}
\alias{dates_na}
\title{Date ranges salted with missing values}
\format{A data frame with 25 rows and 6 variables:
\describe{
  \item{\code{ye}}{integer.}
  \item{\code{yl}}{integer.}
  \item{\code{me}}{integer.}
  \item{\code{ml}}{integer.}
  \item{\code{de}}{integer.}
  \item{\code{dl}}{integer.}
}}
\usage{
dates_na
}
\description{
Date ranges salted with missing values
}
\keyword{datasets}

