% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_monday}
\alias{is_monday}
\title{Predicate function}
\usage{
is_monday(y, m, d)
}
\arguments{
\item{y}{Integer. Candidate year.}

\item{m}{Integer. Candidate month.}

\item{d}{Integer. Candidate day.}
}
\value{
\code{TRUE} if the date is valid, and \code{FALSE} if it is not and
  should be resampled.
}
\description{
An example predicate function that can be passed to \link{\code{sample_date}}
so that it will return only those dates that fall on Mondays.
}

