% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicate}
\alias{is_monday}
\alias{null_predicate}
\alias{predicate}
\title{Predicate functions}
\usage{
is_monday(y, m, d)

null_predicate(y, m, d)
}
\arguments{
\item{y}{Integer. Candidate year.}

\item{m}{Integer. Candidate month.}

\item{d}{Integer. Candidate day.}
}
\value{
\code{TRUE} if the date is valid, and \code{FALSE} if it is not and
  should be resampled.
}
\description{
Predicate functions must take a year, month, and day integer as its
parameters, and return TRUE if the date is valid. Allows one to construct
complex date restrictions, such as only generating dates that are Mondays,
etc.
}
\section{Functions}{
\itemize{
\item \code{is_monday}: An example predicate function that can be passed to
\code{\link{sample_date}} so that it will return only those dates that fall
on Mondays.

\item \code{null_predicate}: Always returns TRUE. This is the default predicate used
by \code{\link{sample_date}}. Used when initially checking values so that
the supplied predicate never rejects a starting condition out-of-hand.
}}

