% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date.R
\name{sample_date}
\alias{sample_date}
\title{Produce random dates within a particular range of possibilities}
\usage{
sample_date(year_min = -9999, year_max = 9999, month_min = 1,
  month_max = 12, day_min = 1, day_max = 31, n)
}
\arguments{
\item{year_min}{Integer. Minimum year (Required)}

\item{year_max}{Integer. Maximum year}

\item{n}{Integer. Number of permuations to return.}

\item{m}{Integer. Month.}

\item{d}{Integer. Day.}
}
\value{
A vector of POSIXct values
}
\description{
Produce random dates within a particular range of possibilities
}
\examples{
sample_date(1930, NA, NA, 5)

sample_date(1930, 2, NA, 5)
}

