% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date.R
\name{sample_date_df}
\alias{sample_date_df}
\title{Return a data frame with these added}
\usage{
sample_date_df(df, y, m, d, n, col_name = "sampled_date")
}
\arguments{
\item{df}{The data frame to augment}

\item{y}{The column of \code{df} containing the year}

\item{m}{The column of \code{df} containing the month}

\item{d}{The column of \code{df} containing the day}

\item{n}{The number of replications}

\item{col_name}{Name of the new column with sampled dates to add to \code{df}}
}
\value{
A data frame
}
\description{
Requires dplyr
}

