% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date.R
\name{sample_date}
\alias{sample_date}
\title{Produce random dates within a particular range of possibilities}
\usage{
sample_date(year_min, year_max = year_min, month_min = 1, month_max = 12,
  day_min = 1, day_max = 31, n, quiet = FALSE)
}
\arguments{
\item{year_min}{Integer. Minimum year (Required)}

\item{year_max}{Integer. Maximum year}

\item{month_min}{Integer. Minimum month}

\item{month_max}{Integer. Maximum month}

\item{day_min}{Integer. Minimum day}

\item{day_max}{Integer. Maximum day}

\item{n}{Integer. Number of permuations to return. (Required)}

\item{quiet}{Boolean. Display messages when regenerating illegal dates?}
}
\value{
A vector of POSIXct values
}
\description{
Produce random dates within a particular range of possibilities
}
\examples{
set.seed(100)
sample_date(1930, NA, NA, 5, quiet = TRUE)

sample_date(1930, 2, NA, 5, quiet = TRUE)
}

