% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date.R
\name{sample_date}
\alias{sample_date}
\title{Produce random dates within a particular range of possibilities}
\usage{
sample_date(y, m, d, n)
}
\arguments{
\item{y}{Integer. Year. (Required)}

\item{m}{Integer. Month.}

\item{d}{Integer. Day.}

\item{n}{Integer. Number of permuations to return.}
}
\value{
A vector of POSIXct values
}
\description{
Produce random dates within a particular range of possibilities
}
\examples{
sample_date(1930, NA, NA, 5)

sample_date(1930, 2, NA, 5)
}

