% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date.R
\name{sample_date}
\alias{sample_date}
\title{Produce random dates within a particular range of possibilities}
\usage{
sample_date(year_min, year_max = year_min, month_min = 1, month_max = 12,
  day_min = 1, day_max = 31, n, .p = null_predicate, quiet = FALSE)
}
\arguments{
\item{year_min}{Integer. Minimum year (Required)}

\item{year_max}{Integer. Maximum year}

\item{month_min}{Integer. Minimum month}

\item{month_max}{Integer. Maximum month}

\item{day_min}{Integer. Minimum day}

\item{day_max}{Integer. Maximum day}

\item{n}{Integer. Number of permuations to return. (Required)}

\item{.p}{An optional predicate function taking a year, month, and day as its
parameters, and returning TRUE if the date is valid. Allows one to
construct complex date restrictions, such as only generating dates that are
Mondays, etc.}

\item{quiet}{Boolean. Display messages when regenerating illegal dates?}
}
\value{
A vector of POSIXct values
}
\description{
Produce random dates within a particular range of possibilities
}
\examples{
set.seed(100)
# Generate random dates within the year 1930
sample_date(1930, n = 5)

# Dates in March 1930
sample_date(1930, month_min = 3, month_max = 3, n = 5)

# Dates between a given range
sample_date(year_min = 1930, year_max = 1931, day_min = 25, day_max = 31, n = 5)

# Use a predicate function to return only Mondays
is_monday <- function(y, m, d) {
 lubridate::wday(lubridate::ymd(paste(y, m, d, sep = "-"), quiet = TRUE)) == 2
}

sample_date(1930, n = 5, .p = is_monday)

}

