% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date_df.R
\name{sample_date_df}
\alias{sample_date_df}
\title{Return a data frame with these added}
\usage{
sample_date_df(df, n, .name = "sampled_date", .id = "replicate",
  quiet = TRUE, ...)
}
\arguments{
\item{df}{The data frame to augment}

\item{n}{Integer. The number of replications}

\item{.name}{Name of the new column with sampled dates to add to \code{df}}

\item{.id}{Name of the new column with iteration ids to add to \code{df}}

\item{quiet}{Boolean.}

\item{...}{Arguments to pass to \code{\link{sample_date}}}
}
\value{
The replicated data frame \code{df} with two new columns.
}
\description{
Requires dplyr
}

